Move HomelabDistribution out of HomelabWasmoOs into its own .kt file. HomelabWasmoOs remains the entrypoint.
