+++ os/distributions/homelab/src/main/kotlin/com/wasmo/distributions/homelab/HomelabDistribution.kt
@@ -1,0 +1,73 @@
+package com.wasmo.distributions.homelab
+
+import com.wasmo.accounts.CookieSecret
+import com.wasmo.accounts.SessionCookieSpec
+import com.wasmo.api.stripe.StripePublishableKey
+import com.wasmo.common.catalog.DevelopmentCatalog
+import com.wasmo.identifiers.Deployment
+import com.wasmo.objectstore.FileSystemObjectStoreAddress
+import com.wasmo.sendemail.postmark.PostmarkCredentials
+import com.wasmo.sendemail.postmark.PostmarkProductionBaseUrl
+import com.wasmo.sql.PostgresqlAddress
+import com.wasmo.sql.ProvisioningDb
+import com.wasmo.stripe.StripeCredentials
+import com.wasmo.wiring.Distribution
+import com.wasmo.wiring.WasmoService
+import dev.zacsweers.metro.createGraphFactory
+import io.ktor.server.engine.EmbeddedServer
+import okhttp3.HttpUrl.Companion.toHttpUrl
+import okio.ByteString.Companion.encodeUtf8
+import okio.Path.Companion.toPath
+import wasmo.sql.SqlDatabase
+
+private val stripePublishableKey = System.getenv("STRIPE_PUBLISHABLE_KEY")
+  ?: error("required env STRIPE_PUBLISHABLE_KEY not set")
+private val stripeSecretKey = System.getenv("STRIPE_SECRET_KEY")
+  ?: error("required env STRIPE_SECRET_KEY not set")
+private val sharedPostgresqlAddress = PostgresqlAddress(
+  user = "postgres",
+  password = "password",
+  hostname = "localhost",
+  databaseName = "wasmo_development",
+  ssl = false,
+)
+
+class HomelabDistribution() : Distribution() {
+  override val osPostgresqlAddress: PostgresqlAddress
+    get() = sharedPostgresqlAddress
+  override val provisioningPostgresqlAddress: PostgresqlAddress
+    get() = sharedPostgresqlAddress
+
+  override fun createService(
+    server: EmbeddedServer<*, *>,
+    provisioningDb: ProvisioningDb,
+    wasmoDb: SqlDatabase,
+  ): WasmoService {
+    val homelabGraphFactory = createGraphFactory<HomelabGraph.Factory>()
+    val serviceGraph = homelabGraphFactory.create(
+      server = server,
+      cookieSecret = CookieSecret("butters".encodeUtf8()),
+      postmarkCredentials = PostmarkCredentials(
+        baseUrl = PostmarkProductionBaseUrl,
+        serverToken = System.getenv("POSTMARK_SERVER_TOKEN") ?: "?",
+      ),
+      stripeCredentials = StripeCredentials(
+        publishableKey = StripePublishableKey(stripePublishableKey),
+        secretKey = stripeSecretKey,
+      ),
+      catalog = DevelopmentCatalog,
+      wasmoDb = wasmoDb,
+      provisioningDb = provisioningDb,
+      postgresqlAddress = sharedPostgresqlAddress,
+      deployment = Deployment(
+        baseUrl = "http://wasmo.localhost:8080/".toHttpUrl(),
+        sendFromEmailAddress = "noreply@wasmo.dev",
+      ),
+      objectStoreAddress = FileSystemObjectStoreAddress(
+        path = System.getProperty("user.home").toPath() / ".wasmo",
+      ),
+      sessionCookieSpec = SessionCookieSpec.Http,
+    )
+    return serviceGraph.wasmoService
+  }
+}
